Added from_sync_state method to block::Tree
This method creates a block::Tree based on SyncState. We also modified the witness field of BlockNode to be Optional, as this value is no longer needed after the node state becomes valid or invalid. This change also eliminates the need for additional witness data when creating block::Tree. Note that our current implementation of archive nodes and synchronization logic is incomplete - archive nodes should maintain complete historical records, while the current SyncState only supports the latest state. We will address this issue in the future