Remove Option from BlockNode Witness field
Not long ago we set the Witness field as Option, but if so, when synchronizing state to other nodes, they cannot perform validation on the data content. This fix is mainly to allow archive nodes to gradually verify all historical records starting from the genesis block. Currently, we don't seem to have a better approach.