Reimplement Default for State
The previous implementation resulted in an empty DAG when State was created using Default. We now add a default block node (via Block::genesis) to the DAG during Default creation. This eliminates the need to handle the special case of an empty DAG every time, and we will modify the code in subsequent commits to remove unnecessary checks.