Add Hasher enum
Previously we used the Hasher trait for hash processing, but by defining all supported hashers explicitly, we can avoid using generics. This commit introduces a Hasher enum and uses Multihash for hash results. (Currently having both the trait and enum with the same name may cause confusion. We plan to standardize on the enum version in the future, but both will coexist temporarily during this transition period.)