Implement Graph
The Graph will be used to manage Atoms. It is essentially an implementation based on a DAG and GHOST. We implemented a basic architecture that supports out-of-order input, checkpoints, a block mechanism, and includes the GHOST logic. This version is only a basic framework and will be further optimized later.