Change conflict detection method
One condition for an Atom to be considered valid is that all outputs of its contained atoms must be unique — i.e., each output's (key, version) pair must not be duplicated (except for mergeable keys). We previously relied on BlockEntry.outputs to enforce this uniqueness, but that was unclear. Switch to using PendingEntry.existing_output to guarantee that all atom