Remove height field from Atom
Remove the height field and derive it from Graph instead. The main reason is that using this field "binds" the Atom to a specific height. For example, when we create an Atom and calculate the VDF, if the Graph's Head moves to what was previously the next block, we would have to recalculate the VDF (since this value changes the VDF input). By removing this field, the Atom can directly use the new Head as its parent (witnesses don't affect VDF input, so atoms can be reassigned after computation completes).