Refactor Command definition to ensure input consumption
The previous Command definition supported "read-only" inputs, but this created a problematic scenario: when all inputs of a Command were already consumed, the Command could still be executed. This would result in the Command being perpetually executable, which is undesirable in any use case.

To address this issue, we have removed the `consumed` field to guarantee that all inputs will be consumed during execution. This change requires modifications to the Graph and Validator design and implementation, which will be addressed in subsequent commits.